exercises7/requirements.txt
fastapi==0.63.0
uvicorn==0.13.3
tensorflow==2.4.1
python-multipart==0.0.5
Pillow==8.1.0
Jinja2==2.11.3
aiofiles==0.6.0
tensorflow-hub==0.11.0